+++ src/pages/Home/pages/MonthSchedulePage/MonthSchedulePage.tsx
@@ -1,7 +1,7 @@
 import MonthlyBudgetSummary from "@pages/Home/next-components/HomeHeader/MonthlyBudgetSummary";
 import ThickDivider from "@components/common/ThickDivider.tsx";
 import CalendarHeader from "@pages/Home/next-components/ScheduleCalendar/CalendarHeader";
-import Calendar from "@pages/Home/next-components/ScheduleCalendar/Calendar";
+import Calendar from "pages/Home/pages/MonthSchedulePage/components/Calendar";
 import moment from "moment/moment";
 import ScheduleList from "@pages/Home/next-components/ScheduleList";
 import useMonthSchedule from "@hooks/home/useMonthSchedule.ts";
