+++ src/pages/Home/pages/MonthSchedulePage/components/Calendar/Calendar.tsx
@@ -1,7 +1,7 @@
 import { DateCalendar, LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterMoment } from "@mui/x-date-pickers/AdapterMoment";
 import moment from "moment";
-import { CalendarBox } from "@pages/Home/next-components/ScheduleCalendar/Calendar/Calendar.styles.ts";
+import { CalendarBox } from "@pages/Home/pages/MonthSchedulePage/components/Calendar/Calendar.styles.ts";
 
 export interface CalendarProps {
   value: string;
